refactor: split virtualized Cell.jsx into per-column renderers
The single Cell.jsx ran every cell render through the same long
if/else chain on column.id and instantiated every hook the dispatcher
needed even when the column only consumed one or two. Splitting into
NameCell, UpdatedAtCell, SizeCell, ShareCell, MenuCell, and
TempDirectoryCell behind a map lookup keeps each renderer focused on
its own data and contexts. A shared useIsInSyncFromSharing hook holds
the row-sync derivation in one place for the three cells that need it.
